Responding to Peter's note about the interface of `HasImpureFinal()`
This function doesn't work for all symbols, and it seemes like putting a
comment on the interface is not a good way to restrict it.  I moved it
to check-do.cpp and put in a check to make sure that it's used
correctly.  I also added a call to `GetAssociationRoot()` to make sure
that it works with all kinds of associations.